backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
websockets==12.0
python-dotenv==1.0.1
httpx==0.27.0
google-genai==1.14.0
chromadb==0.5.3
